# 🏗 Optimize linting during PR builds