--- a/.eslintignore
+++ b/.eslintignore
@@ -32,6 +32,8 @@ build-system/babel-plugins/**/fixtures/**/*.js
 build-system/babel-plugins/**/fixtures/**/*.mjs
 build-system/tasks/make-extension/template/**/*
 build-system/tasks/visual-diff/snippets/*.js
+examples/amp-script/todomvc.ssr.js
+examples/amp-script/vue-todomvc.js
 extensions/amp-a4a/0.1/test/testdata/**
 src/purifier/noop.js
 testing/local-amp-chrome-extension/**