--- a/build-system/pr-check/checks.js
+++ b/build-system/pr-check/checks.js
@@ -70,7 +70,9 @@ async function prBuildWorkflow() {
     timedExecOrDie('amp validate-html-fixtures');
   }
 
-  if (buildTargetsInclude(Targets.LINT)) {
+  if (buildTargetsInclude(Targets.LINT_RULES)) {
+    timedExecOrDie('amp lint');
+  } else if (buildTargetsInclude(Targets.LINT)) {
     timedExecOrDie('amp lint --local_changes');
   }