--- a/build-system/tasks/lint.js
+++ b/build-system/tasks/lint.js
@@ -17,7 +17,6 @@
 
 const argv = require('minimist')(process.argv.slice(2));
 const fs = require('fs');
-const globby = require('globby');
 const {
   log,
   logLocalDev,
@@ -25,11 +24,8 @@ const {
   logOnSameLineLocalDev,
 } = require('../common/logging');
 const {cyan, green, red, yellow} = require('../common/colors');
-const {default: ignore} = require('ignore');
 const {ESLint} = require('eslint');
 const {getFilesToCheck} = require('../common/utils');
-const {gitDiffNameOnlyMain} = require('../common/git');
-const {isCiBuild} = require('../common/ci');
 const {lintGlobs} = require('../test-configs/config');
 
 /** @type {ESLint.Options} */
@@ -134,39 +130,19 @@ function summarizeResults(results, fixedFiles) {
 }
 
 /**
- * Computes the set of files to lint based on command line args / changed files.
- * - If lint rules were changed or if packages were upgraded, lint all files.
- * - Otherwise, use the `getFilesToCheck()` util to parse command line args.
- * @return {!Array<string>}
- */
-function getFilesToLint() {
-  const options = {gitignore: true};
-  const filesChanged = gitDiffNameOnlyMain();
-  const ruleChangeOrPackageUpgrade = filesChanged.some(
-    (file) => file.endsWith('.eslintrc.js') || file == 'package.json'
-  );
-  const overrideLocalChanges =
-    isCiBuild() && argv.local_changes && ruleChangeOrPackageUpgrade;
-  if (overrideLocalChanges) {
-    log(
-      green('INFO:'),
-      'Detected changes to lint rules / packages. Linting all files...'
-    );
-    return ignore().filter(globby.sync(lintGlobs, options));
-  }
-  return getFilesToCheck(lintGlobs, options, '.eslintignore');
-}
-
-/**
  * Checks files for formatting (and optionally fixes them) with Eslint.
  * Explicitly makes sure the API doesn't check files in `.eslintignore`.
  */
 async function lint() {
-  const filesToLint = getFilesToLint();
-  if (filesToLint.length == 0) {
+  const filesToCheck = getFilesToCheck(
+    lintGlobs,
+    {gitignore: true},
+    '.eslintignore'
+  );
+  if (filesToCheck.length == 0) {
     return;
   }
-  await runLinter(filesToLint);
+  await runLinter(filesToCheck);
 }
 
 module.exports = {