@rsimha 2021-05-26
> Perhaps we can run a full lint based on a build target which just checks for linter related files (package.json .eslintrc, etc).

Excellent idea that greatly simplifies the logic! Done.